[B][C][comments] export comment routes to their own module
